Move rainforest_eagle coordinator to separate module
