+++ homeassistant/components/rainforest_eagle/__init__.py
@@ -6,15 +6,15 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 
-from . import data
 from .const import DOMAIN
+from .coordinator import EagleDataCoordinator
 
 PLATFORMS = [Platform.SENSOR]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Rainforest Eagle from a config entry."""
-    coordinator = data.EagleDataCoordinator(hass, entry)
+    coordinator = EagleDataCoordinator(hass, entry)
     await coordinator.async_config_entry_first_refresh()
     hass.data.setdefault(DOMAIN, {})[entry.entry_id] = coordinator
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
